requirements.txt
poppler-utils
streamlit
easyocr
pdf2image
pillow
numpy
